Fix styles, constant 분리 - #12
