+++ src/constants/Admin.constants.ts
@@ -1,0 +1,26 @@
+export const PRODUCT_NAMES = [
+    {
+      id: 1,
+      value: "id",
+    },
+    {
+      id: 2,
+      value: "이름",
+    },
+    {
+      id: 3,
+      value: "종류",
+    },
+    {
+      id: 4,
+      value: "기간/횟수",
+    },
+    {
+      id: 5,
+      value: "가격",
+    },
+    {
+      id: 6,
+      value: "상품수정",
+    },
+  ];
